feat:抽离表单中的匹配器，新增MatchersPanel
